added baseclass for BinClassificator

+++ vsCodeCPPTest/include/data_point.h
@@ -5,12 +5,12 @@
 
 class DataPoint{
     public:
-        DataPoint(int o_class,int x, int y);
+        DataPoint(double o_class,double x, double y);
 //        DataPoint(const DataPoint& obj);
 //        DataPoint(DataPoint&& obj);
         ~DataPoint()=default;
-        const int original_class;
-        int calculated_class;
+        const double original_class;
+        double calculated_class;
         std::vector<double> features;
 };
